requirements.txt
Expedition-Python
pydantic
pyside6
numpy
pandas
pyinstaller
